requirements.txt
black==22.10.0
Django==4.1.4
django-admin-sortable==2.3
django-bootstrap4==22.3
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
ipython==8.7.0
openpyxl==3.0.10
Pillow==9.3.0
psycopg2==2.9.5
requests==2.28.1
redis==4.4
sentry-sdk==1.11.1
sorl-thumbnail==12.9.0
uWSGI==2.0.21
pydantic==2.5.3
django-ninja==1.2.2
weasyprint==60.2
pydyf==0.10.0
mysql.connector
numpy
reportlab
matplotlib
pdfkit